Boundary subclasses and build dependencies
They are not in use yet but we've added a PlaneBoundary class as a
derived class of a Boundary interface base class. It does the same thing
as the old Boundary class. Additionally, we've added a ContainerBoundary
class which is meant to act as a box. It is implemented but not tested.

Whilst doing these changes I came across a crash which I think is due to
the lack of proper dependency tracking between files being compiled. The
Simulation object file was not being re-compiled when Boundary.hh
changed, which it should have been. So after a little research we've
introduced a use of the cpp command to generate the dependencies for the
.cc file we're about to compile. We then pass them to the redo-ifchange
command to make sure everything is ready to go.

The result that cpp -MM returns are not completely ideal. They are
formatted as a multi-line makefile target so we strip out the object
file and colon at the start, then the backslashes and then the new
lines. All with sed.