Introduced logStream and Validator
Added a logStream member to the Simulator class so that we have easy
access to standard out or later on a log file. Should wrap it in DEBUG
statements probably so that it doesn't slow performance but that is
hardly a concern at the moment.

Additionally, we've added a Validator class to the testsuite which
checks over the particles for nan-values and exits the simulation
process if any are found. This makes it much easier to track down the
start of the nans in the logs as otherwise you get pages of nan values
to search through before you see the initial culprit. Now we exit as
soon as one is spotted. We don't exit in a very memory safe way at this
point.

We have also moved the zero-ing of the particle forces to the start of
the simulation step so that the information is there for inspection
between steps. In particular we'd like to draw it.