Switched to Python test suite
We're reaching a point where we want to create and test different sets
of values for the simulations. In order to make this easy it would be
good to have a configuration file format.

However it is irritatingly painful to maintain such a format even if you
use xml/json/yaml and to parse it in C++. So instead we're opting to add
a Python interface which uses the ctypes module to load up our test
program has a library and pass in structured data with the appropriate
config values.

The loading and running has been abstracted into a llyr python module so
all we have to do is create the appropriate instance of the InputData
class and pass that in.

The .do files have been updated accordingly.