Added basic emitter support
Emitter based class and PointEmitter derived class. Results in highly
unstable behaviour unfortunately. Mostly because we make no effort to
check if it is reasonble to create a new particle at a particular
location. ie. will it sit within the kernel of another particle? When it
does is within one there are huge forces applied to both to resolve the
situation which leads to badness.