Create offset grid of particles in intialisation
Instead of randomly placing particles within a block, we create them in
an ordered offset grid. This gives us a much more stable start to the
simulation as the particles fall to the bottom of the container.

Unfortunately when they hit the bottom of the container they explode in
a terribly unstable way. Next thing to look at is a more stable
integration scheme.