Account for the particle's own mass in its density
We were not comparing particles against themselves in the main density
loop due to mistaken concerns over divide by zeros or something.

It makes sense to take a particle's own weight into account when
calculating the density at that point and so we do it and it seems to
greatly stabilise our current setup.