Switch to ParticleData and Midpoint Method
Firstly, we've changed from the array of Particle structs to a single
ParticleData struct that contains arrays of data for each attribute
(position, velocity and mass.)

This allows us to more easier separate out the individual attributes and
pass them around where needed. This is makes it easier to implement the
mid point method as we do calculations with particular attributes whilst
substituting other and so don't want to be passing all the Particle
structs around with special flags to use other arrays, it is better to
split the attribute arrays out and pass only those that are needed.

As a result, most of the code has been tweaked to go from p->pos type
accessors to data.position[i] type accessors.

Secondly, we've switched to the mid-point method. So we've split the
simulator class down somewhat to collaborative objects that can be
called multiple times to calculate the forces and movements needed at
different stages of the method.

The particle data change greatly eases this.

Further improvements should be made for const-correctness and re-use of
memory. We also lose the force visualisation as at the moment the forces
are not kept around between steps to draw.