Added support for multiple initial blocks of particles
Requires a small tweak to the setup C++ and then some juggling of types
in the Python.

In order to make the API kinder we've switched to a create function that
does a bit of work under the hood before returning an InputData
instance, rather than using the InputData contructor directly and
requiring the client to do the extra work.