Added BoxBoundary type and display
We've added a new BoxBoundary class and hardcoded an instance of it
into our setup for the moment.

We've also added a BoxDisplay type to draw our container and the box
boundary and a MultiDisplay type that can take a vector of other
displays and draw them. This allows us to smoothly display the
particles, container and box without disturbing the rest of the setup
code.

We've added a simple collision test case that was used during the box
boundary development. It places a small number of particles directly
above the box and lets them collide.

Finally, we tell the compiler to show all warnings to help keep us in
the loop.