Add headless mode and change viewer to read output files
In order to free us of the glut display rates and have more control over
steps and substeps we change to using a headless mode where the
simulator runs without any display.

We change the viewer and libv.cc to load up the output files from the
simulation and display them.