Bug Fix: Initialise simulation array values
For some reason after the "headless" refactor we were left particularly
prone to a bug where we were using uninitialised arrays. This was of
particular issue with the forces where we only add to the array instead
of initially setting the value and then adding to it.

The best course of action is probably debatable. Perhaps changing our
first += to an = would have been best. This is erring on the safe side
at the moment.