Implement particle bucketing
We implement a Grid class which has the concept of dividng the bounding
box of the particles up into cells, each of width "smoothing distance".

Each cell can store particle ids and we create the grid and populate the
cells at the start of each force evaluation. Where we can we pass in the
known particle bounds otherwise we have to recalculate it if we don't
have it available.

We expand the bounds by 1.01 * "smoothing distance" prior to creating the
grid to make sure we don't have any particles in the outside border of
cells. This means we can generically process each cell, checking its
surrounding cells, without worrying about overrunning the grid.

We introduce a ParticleFactory so with can create Particle instances
with valid bounds.

Finally, in the ForceEvaluator we use the grid so that we don't have to
iterate over n^2 particle tests. We iterate over all the cells in the
grid and for each one, iterate over a set of its neighbours and for each
pairing we iterate over all the particles in each cell and do the
standard comparisons.

We add two checks, first if we're comparing a cell against itself
(i==0&&j==0) we do what we used to which is to only compare particle p
and particle q if p-index is greater than or equal to q-index. This
means we don't double compare particles.

Second, for a particle, out of the nine cell block around it, we
actually only compare it against the following cells:

     i ->
   j -------
   | | | | |
   v -------
     | |x|x|
     -------
     |x|x|x|
     -------

This is because we don't want to double compare cells and as we iterate
over the grid, this pattern should see only comparing particles against
each other once. As a result, we only iterate for i=-1,0,1 and j=0,1 and
we skip the case where i=-1 && j=0 to avoid the middle left cell
comparison.