Switched to yaml config file instead of inline python code
The numerous run configurations in cmd-llyr were become unwieldy and
annoying and I wasn't updating them properly. Better to switch to a
config file which can be saved out to the simulation directory.

It is possibly better to try to create a super lean python script which
can act as a config file, but we're going this way for the moment.

Ideal runs will also be saved with their git SHA1 to reflect the current
state of the code.